features

+++ report.py
@@ -312,7 +312,7 @@ class GetStats:
             if 'data' in data and 'result' in data['data']:
                 values = data['data']['result'][0]['values']
 
-                int_values = [int(value[1]) for value in values]
+                int_values = [float(value[1]) for value in values]
                 self.gpu_eff = sum(int_values) / len(int_values) if int_values else 0
                 #print(f"gpu-usage: {self.gpu_eff}"))
             else:
